Split net.py into seperate files for seperate classes.
