+++ net/__init__.py
@@ -1,3 +1,2 @@
-from net import PiNet
-from net import ArduinoNet
-from net import PiNet
+from pinet import PiNet
+from arduinonet import ArduinoNet
