+++ net/arduinoneterror.py
@@ -1,0 +1,4 @@
+""" The standard exception type to be thrown by instances of ArduinoNet at runtime.
+"""
+class ArduinoNetError(Exception):
+	pass
