+++ net/pineterror.py
@@ -1,0 +1,4 @@
+""" The standard exception type to be thrown by instances of PiNet at runtime.
+"""
+class PiNetError(Exception):
+	pass
